requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
WTForms==3.1.2
python-dotenv==1.0.1
Werkzeug==3.0.3
email-validator==2.1.1
gunicorn
